Results api refactor (#85)
* Make get instead of post requests

* Fixing stash errors

* more merge fixes

* Make get instead of post requests

* Fixing stash errors

* more merge fixes

* Finish up impacted function signatures

* Error handle road user counts route

+++ application/cloud_api.py
@@ -150,7 +150,7 @@ class CloudWizard:
 
         homography = r.json()['homography']
         if file_path:
-            path = os.path.join(file_path, 'homography', 'homography.txt')
+            path = os.path.join(file_path, 'homography.txt')
             np.savetxt(path, np.array(homography))
         return (True, None, homography)
 
@@ -221,7 +221,7 @@ class CloudWizard:
         print "Response Text: {}".format(r.text)
         return self.parse_error(r)
 
-    def getTestConfig(self, identifier, test_flag, project_path):
+    def getTestConfig(self, identifier, test_flag, file_path):
         print "getTestConfig called with identifier = {} and test_flag = {}".format(identifier,test_flag)
 
         payload = {
@@ -230,13 +230,9 @@ class CloudWizard:
         }
 
         if test_flag == 'feature':
-            if not os.path.exists(os.path.join(project_path, 'feature_video')):
-                os.mkdir(os.path.join(project_path, 'feature_video'))
-            path = os.path.join(project_path, 'feature_video', 'feature_video.mp4')
+            path = os.path.join(file_path, 'feature_video.mp4')
         elif test_flag == 'object':
-            if not os.path.exists(os.path.join(project_path, 'object_video')):
-                os.mkdir(os.path.join(project_path, 'object_video'))
-            path = os.path.join(project_path, 'object_video', 'object_video.mp4')
+            path = os.path.join(file_path, 'object_video.mp4')
         else:
             print "ERROR: Invalid flag"
             return (False, 'Invalid test flag: '+str(test_flag))
@@ -395,25 +391,30 @@ class CloudWizard:
 # Results Functions
 ###############################################################################
 
-    def results(self, identifier, ttc_threshold = None, vehicle_only = None, speed_limit = None):
+    def results(self, identifier, project_path, ttc_threshold = None, vehicle_only = None, speed_limit = None):
         print "results called with identifier = {}, ttc_threshold = {}, vehicle_only = {}, and speed_limit= {}"\
                 .format(identifier, ttc_threshold, vehicle_only, speed_limit)
 
         # sync calls
-        s, err = self.roadUserCounts(identifier)
+        s, err = self.roadUserCounts(identifier,\
+                    os.path.join(project_path, 'results'))
         if not s:
             return (s, err)
 
-        s, err = self.speedDistribution(identifier, speed_limit, vehicle_only)
+        s, err = self.speedDistribution(identifier,\
+                    os.path.join(project_path, 'results'),\
+                    speed_limit, vehicle_only)
         if not s:
             return (s, err)
 
-        s, err = self.makeReport(identifier)
+        s, err = self.makeReport(identifier,\
+                    os.path.join(project_path, 'results'))
         if not s:
             return (s, err)
 
         # async calls
-        s, err = self.highlightVideo(identifier, ttc_threshold, vehicle_only)
+        s, err = self.highlightVideo(identifier,\
+                    ttc_threshold, vehicle_only)
         if not s:
             return (s, err)
 
@@ -444,16 +445,45 @@ class CloudWizard:
         }
 
         try:
-            r = requests.post(self.server_addr + 'highlightVideo', data = payload)
+            r = requests.post(self.server_addr + 'highlightVideo', data = payload, streams = True)
         except requests.exceptions.ConnectionError as e:
             print('Connection is offline')
             return (False, 'Connection to server "{}" is offline'.format(self.server_addr))
 
-        print "Status Code: {}".format(r.status_code)
         print "Response Text: {}".format(r.text)
+        print "Status Code: {}".format(r.status_code)
         return self.parse_error(r)
 
-    def makeReport(self, identifier):
+    def getHighlightVideo(self, identifier, file_path):
+        print "getHighlightVideo called with identifier = {}".format(identifier)
+
+        payload = {
+            'identifier': identifier
+        }
+
+        try:
+            r = requests.get(self.server_addr + 'highlightVideo', params = payload, streams = True)
+        except requests.exceptions.ConnectionError as e:
+            print('Connection is offline')
+            return (False, 'Connection to server "{}" is offline'.format(self.server_addr))
+
+        success, err = self.parse_error(r)
+        if not success:
+            return (success, err)
+
+        path = os.path.join(file_path, 'highlight.mp4')
+        if os.path.exists(path):
+            os.remove(path)
+
+        with open(path, 'wb') as f:
+            print('Dumping "{0}"...'.format(path))
+            for chunk in r.iter_content(chunk_size=2048):
+                if chunk:
+                    f.write(chunk)
+
+        return (True, None)
+
+    def makeReport(self, identifier, file_path):
         print "makeReport called with identifier = {}".format(identifier)
 
         payload = {
@@ -461,24 +491,34 @@ class CloudWizard:
         }
 
         try:
-            r = requests.post(self.server_addr + 'makeReport', data = payload)
+            r = requests.get(self.server_addr + 'makeReport', params  = payload, stream = True)
         except requests.exceptions.ConnectionError as e:
             print('Connection is offline')
             return (False, 'Connection to server "{}" is offline'.format(self.server_addr))
 
-        print "Status Code: {}".format(r.status_code)
-        print "Response Text: {}".format(r.text)
-        return self.parse_error(r)
+        success, err = self.parse_error(r)
+        if not success:
+            return (success, err)
 
-    def retrieveResults(self, identifier, project_path):
+        path = os.path.join(file_path, 'santosreport.pdf')
+        if os.path.exists(path):
+            os.remove(path)
+
+        with open(path, 'wb') as f:
+            print('Dumping "{0}"...'.format(path))
+            for chunk in r.iter_content(chunk_size=2048):
+                if chunk:
+                    f.write(chunk)
+
+        print "Status Code: {}".format(r.status_code)
+        return (True, None)
+
+    def retrieveResults(self, identifier, file_path):
         print "retrieveResults called with identifier = {}".format(identifier)
 
         payload = {
             'identifier': identifier,
         }
-        path = os.path.join(project_path, 'results', 'results.zip')
-        if os.path.exists(path):
-            os.remove(path)
 
         try:
             r = requests.get(self.server_addr + 'retrieveResults', params = payload, stream=True)
@@ -490,6 +530,10 @@ class CloudWizard:
         if not success:
             return (success, err)
 
+        path = os.path.join(file_path, 'results.zip')
+        if os.path.exists(path):
+            os.remove(path)
+
         with open(path, 'wb') as f:
             print('Dumping "{0}"...'.format(path))
             for chunk in r.iter_content(chunk_size=2048):
@@ -499,7 +543,7 @@ class CloudWizard:
         print "Status Code: {}".format(r.status_code)
         return (True, None)
 
-    def roadUserCounts(self, identifier):
+    def roadUserCounts(self, identifier, file_path):
         print "roadUserCounts called with identifier = {}".format(identifier)
 
         payload = {
@@ -507,16 +551,29 @@ class CloudWizard:
         }
 
         try:
-            r = requests.post(self.server_addr + 'roadUserCounts', data = payload)
+            r = requests.get(self.server_addr + 'roadUserCounts', params = payload, stream=True)
         except requests.exceptions.ConnectionError as e:
             print('Connection is offline')
             return (False, 'Connection to server "{}" is offline'.format(self.server_addr))
 
-        print "Status Code: {}".format(r.status_code)
-        print "Response Text: {}".format(r.text)
-        return self.parse_error(r)
+        success, err = self.parse_error(r)
+        if not success:
+            return (success, err)
 
-    def speedDistribution(self, identifier, speed_limit = None, vehicle_only = None):
+        path = os.path.join(file_path, 'road_user_icon_counts.jpg')
+        if os.path.exists(path):
+            os.remove(path)
+
+        with open(path, 'wb') as f:
+            print('Dumping "{0}"...'.format(path))
+            for chunk in r.iter_content(chunk_size=2048):
+                if chunk:
+                    f.write(chunk)
+
+        print "Status Code: {}".format(r.status_code)
+        return (True, None)
+
+    def speedDistribution(self, identifier, file_path, speed_limit = None, vehicle_only = None):
         print "speedDistribution called with identifier = {}, speed_limit = {} and vehicle_only = {}"\
                 .format(identifier, speed_limit, vehicle_only)
 
@@ -527,16 +584,27 @@ class CloudWizard:
         }
 
         try:
-            r = requests.post(self.server_addr + 'speedDistribution', data = payload)
+            r = requests.get(self.server_addr + 'speedDistribution', params = payload, stream=True)
         except requests.exceptions.ConnectionError as e:
             print('Connection is offline')
             return (False, 'Connection to server "{}" is offline'.format(self.server_addr))
 
+        success, err = self.parse_error(r)
+        if not success:
+            return (success, err)
+
+        path = os.path.join(file_path, 'velocityPDF.jpg')
+        if os.path.exists(path):
+            os.remove(path)
+
+        with open(path, 'wb') as f:
+            print('Dumping "{0}"...'.format(path))
+            for chunk in r.iter_content(chunk_size=2048):
+                if chunk:
+                    f.write(chunk)
+
         print "Status Code: {}".format(r.status_code)
-        print "Response Text: {}".format(r.text)
-        return self.parse_error(r)
-
-
+        return (True, None)
 
 ###############################################################################
 # Helper Methods
